Med-DigestPro/requirements.txt
flask
requests
beautifulsoup4
google-generativeai
python-dotenv
flask_sqlalchemy
fpdf
feedparser